requirements.txt
llama-index
python-dotenv
html2text
ipykernel
llama-index-llms-openai
llama-index-readers-web